requirements.txt
# Base requirements
Pillow>=8.4.0
fastparquet>=0.8.1
imgaug>=0.4.0
loguru>=0.5.3
matplotlib>=3.5
minio>=7.1.7
numpy>=1.21
nvidia-ml-py3>=7.352.0
pandas>=1.3
poetry>=1.1.13
protobuf==3.20
pyarrow>=7.0.0
pybboxes==0.0.2
pymongo[srv]>=4.0.2
python-color-transfer==0.1.0b0
python-dotenv>=0.19.1
ray>=1.11.0
requests>=2.27.1
schedule>=1.1.0
seaborn>=0.11
sphinx>=5.0.1
sphinx_rtd_theme>=1.0.0
tqdm>=4.62.3
wandb>=0.12.15
sqlalchemy>=1.4.36

# API requirements
fastapi>=0.75.2
fastapi_jwt_auth>=0.5.0
pydantic>=1.8.2
rayim==1.3.0
uvicorn[standard]>=0.17.6

# YOLOv5 requirements
PyYAML>=5.3.1
opencv-python>=4.1.2
python-multipart>=0.0.5
scipy>=1.4.1
tensorboard>=1.15
torch>=1.7.0
torchvision>=0.8.1
